passing hm tests up to g_overload_13
All modes too...
Drop the special resolve() pass; move needed code into the main HM pass.
Always run HM to resolve fields, even for GCP-only runs.
unify will attempt to resolve fields, sorta like a version of 'find()' on labels.
Back to alpha sorting fields in structs, but also keep bi-jection property.  Unsorted fields are an error.
Make a fcn for common missing-field errors.
Restructure _rflds; arg_meet changing types in Pair/Triple needs to revisit Fields.
Rename DO_AMBI as HM_AMBI; use Field.FIELDS as the ambiguous worklist.

+++ src/main/java/com/cliffc/aa/util/Util.java
@@ -35,7 +35,7 @@ public class Util {
     long sum=0, ssum=0;
     for( int i=0; i<ps._len; i++ ) {
       System.out.print(""+i+":"+ps.at(i)+", ");
-      long l = (long)ps.at(i);
+      long l = ps.at(i);
       sum  += l;
       ssum += l*(i+1);
     }
